+++ app/pages/AboutPage/index.ts
@@ -1,0 +1,1 @@
+export * from './AboutPage';
