--- app/pages/Covid/index.ts
@@ -1,1 +1,0 @@
-export * from './Covid';
