+++ app/pages/CovidPage/CovidPage.module.scss
