+++ app/pages/CovidPage/components/CovidHeader/CovidHeader.jsx
