+++ app/pages/CovidPage/components/CovidHeader/CovidHeader.module.scss
