+++ app/pages/CovidPage/components/CovidHeader/index.js
