+++ app/pages/CovidPage/components/FoodList/FoodList.jsx
