+++ app/pages/CovidPage/components/FoodList/FoodList.module.scss
