+++ app/pages/CovidPage/components/FoodList/index.js
