+++ app/pages/CovidPage/components/FoodMap/FoodMap.jsx
