+++ app/pages/CovidPage/components/FoodMap/FoodMap.module.scss
