+++ app/pages/CovidPage/components/FoodMap/index.js
