+++ app/pages/CovidPage/components/GoogleDocEmbed/GoogleDocEmbed.module.scss
