+++ app/pages/CovidPage/components/SimpleGuide/index.js
