+++ app/pages/CovidPage/index.ts
@@ -1,0 +1,1 @@
+export * from './CovidPage';
